Moved some files and added a proper main file.
