+++ haros/plugin_manager.py
